requirements.txt
# Requires Python >= 3.10 and <= 3.12
beautifulsoup4==4.14.3
pillow==12.2.0
requests==2.33.1
tqdm==4.67.3